Rc 1.8.0

+++ 01_Data/src/layers/sequelize/util.ts
@@ -1,7 +1,6 @@
-// Copyright (c) 2023 S44, LLC
-// Copyright Contributors to the CitrineOS Project
+// SPDX-FileCopyrightText: 2025 Contributors to the CitrineOS Project
 //
-// SPDX-License-Identifier: Apache 2.0
+// SPDX-License-Identifier: Apache-2.0
 
 import { type BootstrapConfig } from '@citrineos/base';
 import { type Dialect } from 'sequelize';
@@ -9,7 +8,6 @@ import { Sequelize } from 'sequelize-typescript';
 import { type ILogObj, Logger } from 'tslog';
 import { ComponentVariable } from './model/DeviceModel/ComponentVariable';
 import {
-  AdditionalInfo,
   Authorization,
   Boot,
   Certificate,
@@ -26,13 +24,14 @@ import {
   Connector,
   EventData,
   Evse,
-  IdToken,
-  IdTokenInfo,
+  EvseType,
   InstalledCertificate,
+  LatestStatusNotification,
   LocalListAuthorization,
   LocalListVersion,
   LocalListVersionAuthorization,
   Location,
+  MessageInfo,
   MeterValue,
   OCPPMessage,
   Reservation,
@@ -41,7 +40,14 @@ import {
   SendLocalList,
   SendLocalListAuthorization,
   ServerNetworkProfile,
+  SetNetworkProfile,
+  StartTransaction,
+  StatusNotification,
+  StopTransaction,
+  Subscription,
+  Tariff,
   Tenant,
+  TenantPartner,
   Transaction,
   TransactionEvent,
   Variable,
@@ -49,15 +55,8 @@ import {
   VariableCharacteristics,
   VariableMonitoring,
   VariableMonitoringStatus,
+  VariableStatus,
 } from '.';
-import { VariableStatus } from './model/DeviceModel';
-import { MessageInfo } from './model/MessageInfo';
-import { Subscription } from './model/Subscription';
-import { Tariff } from './model/Tariff';
-import { IdTokenAdditionalInfo } from './model/Authorization/IdTokenAdditionalInfo';
-import { SetNetworkProfile, StatusNotification } from './model/Location';
-import { LatestStatusNotification } from './model/Location/LatestStatusNotification';
-import { StartTransaction, StopTransaction } from './model/TransactionEvent';
 
 export class DefaultSequelizeInstance {
   /**
@@ -108,15 +107,26 @@ export class DefaultSequelizeInstance {
         }
       }
     }
+    this.logger.info(`Sequelize initialized: ${JSON.stringify(this.instance?.config || {})}`);
   }
 
-  private static async syncDb() {
-    if (this.config.database.alter) {
-      await this.instance!.sync({ alter: true });
-      this.logger.info('Database altered');
-    } else if (this.config.database.sync) {
-      await this.instance!.sync({ force: true });
-      this.logger.info('Database synchronized');
+  private static async syncDb(): Promise<void> {
+    if (this.config.database.sync) {
+      const alter = this.config.database.alter;
+      const force = this.config.database.force;
+      if (force) {
+        this.logger.info('Database force synchronizing');
+        await this.instance!.sync({ force: true });
+        this.logger.info('Database force synchronized');
+      } else if (alter) {
+        this.logger.info('Database altering');
+        await this.instance!.sync({ alter: true });
+        this.logger.info('Database altered');
+      } else {
+        this.logger.info('Database synchronizing');
+        await this.instance!.sync();
+        this.logger.info('Database synchronized');
+      }
     }
   }
 
@@ -129,7 +139,6 @@ export class DefaultSequelizeInstance {
       username: this.config.database.username,
       password: this.config.database.password,
       models: [
-        AdditionalInfo,
         Authorization,
         Boot,
         Certificate,
@@ -147,10 +156,8 @@ export class DefaultSequelizeInstance {
         CompositeSchedule,
         Connector,
         Evse,
+        EvseType,
         EventData,
-        IdToken,
-        IdTokenAdditionalInfo,
-        IdTokenInfo,
         Location,
         MeterValue,
         MessageInfo,
@@ -180,7 +187,9 @@ export class DefaultSequelizeInstance {
         SendLocalList,
         SendLocalListAuthorization,
         Tenant,
+        TenantPartner,
       ],
+      pool: this.config.database.pool,
       logging: (_sql: string, _timing?: number) => {},
     });
   }
